requirements.txt
# Core dependencies
streamlit==1.43.2
neo4j==5.15.0
py2neo==2021.2.4
networkx==3.2.1

# ML and NLP - FIXED COMPATIBLE VERSIONS
numpy==1.24.4
scipy==1.10.1
torch==2.0.1+cpu
torchvision==0.15.2+cpu 
torchaudio==2.0.2+cpu
scikit-learn==1.3.2
sentence-transformers==2.2.2
transformers==4.30.2

# Computer Vision - HEADLESS VERSION (no GUI dependencies)
opencv-python-headless==4.7.0.72
easyocr==1.7.0
Pillow>=9.0.0

# NLP
spacy==3.7.2
nltk==3.9.1
huggingface-hub==0.16.4

# LangChain ecosystem
langchain==0.3.20
langchain-core==0.3.44
langchain-groq==0.2.5
langchain-text-splitters==0.3.6

# Data processing
pandas==2.2.3
PyMuPDF==1.25.3

# Visualization
plotly==6.0.0
altair==5.5.0

# Utilities
python-dotenv==1.0.1
pydantic==2.10.6
python-dateutil==2.9.0.post0

# Communication
sendgrid==6.11.0
fpdf==1.7.2

# Additional dependencies
groq==0.19.0
streamlit-option-menu==0.3.6
streamlit-agraph==0.0.45
